% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{as_pfc}
\alias{as_pfc}
\title{Make a \code{pfc} object from a \code{phylo} object}
\usage{
as_pfc(phy)
}
\arguments{
\item{phy}{A phylogenetic tree in \code{ape::phylo}
format}
}
\value{
a \code{pfc}
}
\description{
Make a \code{pfc} object from a \code{phylo} object
}
\examples{
as_pfc(ape::rtree(100))
}
