% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregion_ids}
\alias{ecoregion_ids}
\title{Ecoregion IDs}
\format{
\subsection{\code{ecoregion_ids}}{

A 'pf' data frame (subclasses \code{tibble}) with 847 rows and 2 columns:
\describe{
\item{ECO_ID}{ID numbers for 847 ecoregions across the world}
\item{ECO_NAME}{The name for the corresponding ecoregion ID}
}
}
}
\source{
\url{https://ecoregions.appspot.com}
}
\usage{
ecoregion_ids
}
\description{
A \code{data.frame} with ecoregion ids and their name. This can be matched to the
ecoregions referred to in the dataset \code{mammal_biogeo}.
}
\keyword{datasets}
