% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_interp.R
\name{make_interp}
\alias{make_interp}
\title{Make a phylogenetic interpolation \code{pfc}}
\usage{
make_interp(x, ...)
}
\arguments{
\item{x}{A \code{pfc} or \code{pf} object}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Make a phylogenetic interpolation \code{pfc}
}
