% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammal_biogeo}
\alias{mammal_biogeo}
\title{Terrestrial Mammal Bioegeography Data with Phylogeny}
\format{
\subsection{\code{mammal_biogeo}}{

A 'pf' data frame (subclasses \code{tibble}) with 10,648 rows and 844 columns:
\describe{
\item{label}{Node labels including species name for the tip labels}
\item{phlo}{The phylogenetic flow column which stores the phylogenetic information}
\item{IUCN_binomial}{Species name used by the IUCN, matches with IUCN range polygons}
\item{body_mass_median}{Mammal species' body mass}
\item{litter_clutch_size}{Mammal species' average clutch size
(# of offspring in a litter)}
\item{activity)}{Mammal species' primary time of activity}
\item{hab_breadth}{...}
\item{volant}{...}
\item{diet_5cat}{...}
\item{range_size_km2}{Mammal species' range size in kilometers squared}
\item{threat}{Mammal species' IUCN threat category}
\item{ecoregion:eco_id}{Each of the 832 columns starting with "ecoregion:"
represents the proportion of the mammal species' range that fall in the ecoregion
with id equal to "eco_id".}
}
}
}
\source{
\url{https://ecoregions.appspot.com}, \url{https://www.iucnredlist.org/resources/spatial-data-download}
}
\usage{
mammal_biogeo
}
\description{
Data on terrestrial mammals biogeographic distributions across the world's ecoregions
joined to a \code{pf} object (for \code{{phyf}})
}
\references{
None yet.
}
\keyword{datasets}
