% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_anc}
\alias{pf_anc}
\title{Return a \code{pfc} with ancestral features}
\usage{
pf_anc(x, replace = 0)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{replace}{Value to replace edge feature with no
ancestral value (such as when the ancestor is the root)}
}
\value{
A new \code{pfc} with the same structure as \code{x}, but with
the features of each edge's ancestors instead
}
\description{
Return a \code{pfc} with ancestral features
}
\examples{
pf_anc(rpfc(100))
}
