% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_as_pf}
\alias{pf_as_pf}
\title{Convert a \code{ape::phylo} object to a \code{pf} object}
\usage{
pf_as_pf(phy)
}
\arguments{
\item{phy}{A \code{phylo} object}
}
\value{
a \code{pf} object with branch lengths as features
}
\description{
Convert a \code{ape::phylo} object to a \code{pf} object
}
\examples{
pf_as_pf(ape::rtree(100))
}
