% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_as_pfc}
\alias{pf_as_pfc}
\title{Make a \code{pfc} object from a \code{phylo} object}
\usage{
pf_as_pfc(x, ...)
}
\arguments{
\item{x}{A phylogenetic tree in \code{ape::phylo}
format}

\item{...}{Arguments passed to or from other methods}
}
\value{
a \code{pfc}
}
\description{
Make a \code{pfc} object from a \code{phylo} object
}
\examples{
pf_as_pfc(ape::rtree(100))
}
