% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_as_pftibble}
\alias{pf_as_pftibble}
\title{Convert pfc into a pftibble}
\usage{
pf_as_pftibble(x)
}
\arguments{
\item{x}{a \code{pfc} object}
}
\value{
A \code{tibble}
}
\description{
Convert pfc into a pftibble
}
\examples{
pf_as_pftibble(rpfc(100))
}
