% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_as_phylo}
\alias{pf_as_phylo}
\title{Convert a \code{pf} or \code{pfc} object to a \code{ape::phylo} object}
\usage{
pf_as_phylo(x, ...)
}
\arguments{
\item{x}{a \code{pf} or \code{pfc} object to be converted}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A \code{phylo} object
}
\description{
This function attempts to convert a \code{pf} or \code{pfc} object
to a phylogenyin \code{phylo} format (from package \code{{ape}}).
It will use the feature as branch lengths, and if edges
have multiple feature values, they will be aggregated
by averaging them. Note that this function can fail if the
\code{pf} or \code{pfc} does not have a tree-like structure. An example
of this would be an 'interaction' \code{pfc} (as generated by
\code{pf_interaction})
}
\examples{
pf_as_phylo(rpfc(100))
}
