% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_as_sparse}
\alias{pf_as_sparse}
\title{Convert a \code{pf} or \code{pfc} object to a sparse matrix
representation.}
\usage{
pf_as_sparse(x, ...)
}
\arguments{
\item{x}{A \code{pfc} or \code{pf} object}

\item{...}{Arguments passed to or from other methods}
}
\value{
A \code{dgCMatrix} object (from the \verb{\{Matrix\} package}). See
\code{Matrix::sparseMatrix()} for details.
}
\description{
Convert a \code{pf} or \code{pfc} object to a sparse matrix
representation.
}
\examples{
pf_as_sparse(rpfc(100))
}
