% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_desc}
\alias{pf_desc}
\title{Return a \code{pfc} with descendent's features}
\usage{
pf_desc(x, replace = 0)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{replace}{Value to replace edge feature with no
descendent values (such as when the edges are the tip edges).}
}
\value{
A new \code{pfc} with the same structure as \code{x}, but with
the features of each edge's descendant instead
}
\description{
Note that in the context of a phylogenetic flow, each element of the flow
only has a single descendent, which is the next edge on the sequence between
the root and the terminal node of the flow in question.
}
\examples{
pf_desc(rpfc(100))
}
