% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_edge_apply}
\alias{pf_edge_apply}
\title{Apply a function along edges within a \code{pfc} object}
\usage{
pf_edge_apply(x, fun, ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{fun}{A function to apply to the edge vectors, which will be passed as
the first argument. Must return a vector the same length and type as the input.}

\item{...}{Any additional arguments to pass to \code{fun}}
}
\value{
A \code{pfc} object with the same structure as \code{x} but with features
replaced with the output of \code{fun}
}
\description{
Features of the returned \code{pfc} will be the output of the function applied
to the original features. Therefore the functionm provided must return a
vector the same length of the input.
}
\examples{
## average features of descendents of each edge as a pfc:
pf_edge_apply(pf_desc(pf_indexes(rpfc(100))), function(x) rep(mean(x), length(x)))
}
