% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_edge_names}
\alias{pf_edge_names}
\title{Extract edge names from \code{pfc} object}
\usage{
pf_edge_names(x)
}
\arguments{
\item{x}{\code{pfc} object}
}
\value{
A character vector of edge names
}
\description{
Extract edge names from \code{pfc} object
}
\examples{
pf_edge_names(rpfc(100))
}
