% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_manip.R
\name{pf_edge_segmentize}
\alias{pf_edge_segmentize}
\title{Segments the edges of a phylogeny by splitting them at particular positions}
\usage{
pf_edge_segmentize(x, edges, positions)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{edges}{A vector of edge names}

\item{positions}{A vector of position to cut the edges (must be same length as \code{edges})}
}
\value{
A \code{pfc} object with edges segmented
}
\description{
Segments the edges of a phylogeny by splitting them at particular positions
}
\examples{
pf_edge_segmentize(rpfc(100), "t1", 0.01)
}
