% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_end_features}
\alias{pf_end_features}
\title{Extract the features of the end edge of each phylogenetic flow}
\usage{
pf_end_features(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A numeric vector of the features of the end edge of each phylogenetic flow
}
\description{
Extract the features of the end edge of each phylogenetic flow
}
\examples{
pf_end_features(rpfc(100))
}
