% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_ends}
\alias{pf_ends}
\title{Returns the end edge of each phylogenetic flow as a two-column tibble
with start and end columns}
\usage{
pf_ends(x, return_names = TRUE)
}
\arguments{
\item{x}{A pfc object}

\item{return_names}{If TRUE, return the start and end node names. If FALSE,
return their indexes.}
}
\value{
A \code{tibble} with start and end nodes
}
\description{
Returns the end edge of each phylogenetic flow as a two-column tibble
with start and end columns
}
\examples{
pf_ends(rpfc(100))
}
