% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_manip.R
\name{pf_epoch_info}
\alias{pf_epoch_info}
\title{Get edges and positions along edges where a set of epoch times intersect}
\usage{
pf_epoch_info(x, times)
}
\arguments{
\item{x}{An object of class pfc}

\item{times}{A vector of epoch times to slice the pfc along}
}
\value{
A tibble with edge labels in the first column and position in the second column
}
\description{
Get edges and positions along edges where a set of epoch times intersect
}
\examples{
pf_epoch_info(rpfc(100), c(1, 2, 3))
}
