% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_flow_cumsum}
\alias{pf_flow_cumsum}
\title{Calculate the cumulative sum of features for each flow}
\usage{
pf_flow_cumsum(x, direction = c("from_root", "to_root"), ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{direction}{Which direction to take the cumulative sum along?
"from_root", the default, goes from the root to the terminal nodes.
"to_root" goes in the opposite direction, from the terminal nodes to
the root.}

\item{...}{Other arguments to pass to or from other methods}
}
\value{
A \code{pfc} object with cumulative sums of features in
place of the original features
}
\description{
Cumulative sum the phylogenetic flow features for each flow
in a \code{pfc}, returning a \code{pfc}.
}
\examples{
pf_flow_cumsum(rpfc(100))
}
