% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_flow_sum}
\alias{pf_flow_sum}
\title{Calculate the sum of features for each flow}
\usage{
pf_flow_sum(x, ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{...}{Other arguments to pass to or from other methods}
}
\value{
A vector equal to the length of \code{x} with the
sum of each flow's features
}
\description{
Sums the phylogenetic flow features for each flow
in a \code{pfc}, returning a vector of values.
}
\examples{
pf_flow_sum(rpfc(100))
}
