% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_indexes}
\alias{pf_indexes}
\title{Replace feature with edge index in a \code{pfc}}
\usage{
pf_indexes(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A new \code{pfc} with the same structure as \code{x}, but with
the features of each edge equal to the edge index (the column number in the
sparse matrix representation)
}
\description{
This is similar to base R's \code{col()} function, it returns a \code{pfc} with features
that are just the index of the edge. Useful for complex indexing procedures
}
\examples{
pf_indexes(rpfc(100))
}
