% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_internal}
\alias{pf_internal}
\alias{pf_internal<-}
\alias{pf_terminal}
\alias{pf_terminal<-}
\title{Extract or assign into the internal or terminal edges of a \code{pfc}}
\usage{
pf_internal(x, ...)

pf_internal(x) <- value

pf_terminal(x, ...)

pf_terminal(x) <- value
}
\arguments{
\item{x}{a \code{pfc} object}

\item{...}{Arguments passed to or from other methods.}

\item{value}{a \code{pfc} or \code{dgCMatrix} with the
same dimensions as \code{pf_internal(x)} or \code{pf_terminal(x)}}
}
\value{
a \code{pfc} object with flows truncated to internal nodes.
}
\description{
\code{pf_internal} extracts or assigns a \code{pfc} or \code{dgCMatrix} into the internal
edges of a \code{pfc}.
}
\details{
\code{pf_terminal} extracts or assigns a \code{pfc} or \code{dgCMatrix} into the terminal
edges of a \code{pfc}.
}
\examples{
pfc1 <- rpfc(100)
# Pagel's lambda transformation:
lambda <- 0.5
lambda_pfc <- pfc1
root2node_lens <- pf_flow_sum(lambda_pfc)
pf_internal(lambda_pfc) <- pf_internal(lambda_pfc) * lambda
pf_terminal(lambda_pfc) <- pf_terminal(lambda_pfc) * (root2node_lens / pf_flow_sum(lambda_pfc))
plot(lambda_pfc)
}
