% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{pf_is_desc}
\alias{pf_is_desc}
\title{Return a logical vector determining if a flow's terminal node is
a descendant of an edge.}
\usage{
pf_is_desc(x, edge)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{edge}{An edge number or name}
}
\value{
A logical vector
}
\description{
Return a logical vector determining if a flow's terminal node is
a descendant of an edge.
}
\examples{
pf_is_desc(rpfc(100), "Node12")
}
