% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_is_empty}
\alias{pf_is_empty}
\title{Test which elements of a pfc are empty}
\usage{
pf_is_empty(x, ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{...}{Further arguments for future extensions. Currently not used.}
}
\value{
A logical vector of the same length as \code{x} with \code{TRUE} where
there are empty flows in \code{x}, \code{FALSE} otherwise
}
\description{
Test which elements of a pfc are empty
}
\examples{
pf_is_empty(rpfc(100))
}
