% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_is_tips}
\alias{pf_is_tips}
\title{Return a logical vector which is \code{TRUE} for the elements
of a \code{pfc} whci reprsent tips of a phylogeny}
\usage{
pf_is_tips(x, ...)
}
\arguments{
\item{x}{A \code{pfc} object.}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Return a logical vector which is \code{TRUE} for the elements
of a \code{pfc} whci reprsent tips of a phylogeny
}
\examples{
pf_is_tips(rpfc(100)) 
}
