% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pf_kronecker}
\alias{pf_kronecker}
\title{Calculate a kronecker product when a \code{pfc} is the
multiplicand or the multiplier.}
\usage{
pf_kronecker(x, y, ...)
}
\arguments{
\item{x}{The kronecker multiplicand.}

\item{y}{The kronecker multiplier.}

\item{...}{Other arguments passed to or from other methods.}
}
\value{
A \code{pfc} object
}
\description{
Calculate a kronecker product when a \code{pfc} is the
multiplicand or the multiplier.
}
\examples{
pf_kronecker(rpfc(20), rpfc(20))
}
