% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_labels}
\alias{pf_labels}
\title{Return a vector of labels for a \code{pfc} object}
\usage{
pf_labels(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A character vector of labels
}
\description{
Return a vector of labels for a \code{pfc} object
}
\examples{
pf_labels(rpfc(100))
}
