% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_mean_edge_features}
\alias{pf_mean_edge_features}
\title{Extract mean edge features from \code{pfc} object}
\usage{
pf_mean_edge_features(x)
}
\arguments{
\item{x}{\code{pfc} object}
}
\value{
A numeric vector of mean edge features
}
\description{
Extract mean edge features from \code{pfc} object
}
\examples{
pf_mean_edge_features(rpfc(100))
}
