% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetic.R
\name{pf_mrca}
\alias{pf_mrca}
\title{Return the edge leading up to the most recent common ancestor of a set
of phylogenetic flows}
\usage{
pf_mrca(x, name = FALSE, ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{name}{Should the edge name be returned? If \code{FALSE}, the default,
the edge number is returned}

\item{...}{Other arguments passed to or from other methods}
}
\value{
A string with the edge name is \code{name = TRUE}, or an integer
index otherwise
}
\description{
Return the edge leading up to the most recent common ancestor of a set
of phylogenetic flows
}
\examples{
require(dplyr)
avonet \%>\%
  filter(label \%in\% c("Platalea_minor", "Pelecanus_occidentalis")) \%>\%
  pull(phlo) \%>\%
  pf_mrca()
}
