% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_nedges}
\alias{pf_nedges}
\title{Return the number of edges in a \code{pfc}}
\usage{
pf_nedges(x)
}
\arguments{
\item{x}{a \code{pfc} object}
}
\value{
The number of edges
}
\description{
Return the number of edges in a \code{pfc}
}
\examples{
pf_nedges(rpfc(100))
}
