% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_nodes}
\alias{pf_nodes}
\title{Get only the (internal) node elements of a \code{pfc}}
\usage{
pf_nodes(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A \code{pfc} object with only internal node elements
}
\description{
Get only the (internal) node elements of a \code{pfc}
}
\examples{
pf_nodes(rpfc(100))
}
