% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_ones}
\alias{pf_ones}
\title{Replace features with ones}
\usage{
pf_ones(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A \code{pfc} object with all feature values replaced with ones
}
\description{
Replace features with ones
}
\examples{
pf_ones(rpfc(100))
}
