% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_path}
\alias{pf_path}
\title{Extract \code{pfp} object from \code{pfc}, the paths of
each flow from root to terminal node.}
\usage{
pf_path(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A \code{pfp} object
}
\description{
Extract \code{pfp} object from \code{pfc}, the paths of
each flow from root to terminal node.
}
\examples{
pf_path(rpfc(100))
}
