% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_phyloflow}
\alias{pf_phyloflow}
\title{Extracts the phylogenetic flow column of an \code{pf} object}
\usage{
pf_phyloflow(x)
}
\arguments{
\item{x}{A \code{pf} object}
}
\value{
A \code{pfc} object (phylogenetic flow collection)
}
\description{
Extracts the phylogenetic flow column of an \code{pf} object
}
