% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_position}
\alias{pf_position}
\title{Replace feature with edge position in flow in a \code{pfc}}
\usage{
pf_position(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A new \code{pfc} with the same structure as \code{x}, but with
the features of each edge equal to the position along the flow as an integer
index
}
\description{
Replace feature with edge position in flow in a \code{pfc}
}
\examples{
pf_position(rpfc(100))
}
