% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pf_row_kron}
\alias{pf_row_kron}
\title{Calculate a rowwise kronecker product when the multiplicand or multiplier
is a \code{pfc}.}
\usage{
pf_row_kron(x, y, ...)
}
\arguments{
\item{x}{The rowwise kronecker multiplicand}

\item{y}{The rowwise kronecker multiplier}

\item{...}{Other arguments passed to or from other methods.}
}
\value{
A \code{pfc} object.
}
\description{
Calculate a rowwise kronecker product when the multiplicand or multiplier
is a \code{pfc}.
}
\examples{
pf_row_kron(rpfc(20), rpfc(20))
}
