% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_scale_flow_sum}
\alias{pf_scale_flow_sum}
\title{Scale the phylogenetic flow features to a constant sum.}
\usage{
pf_scale_flow_sum(x, scale_to = 1)
}
\arguments{
\item{x}{a \code{pfc} object}

\item{scale_to}{The value to scale the sums to. Default is 1.}
}
\value{
a scaled \code{pfc} object
}
\description{
Can be used to standardise branch lengths to reasonable values
}
\examples{
pf_scale_flow_sum(rpfc(100))
}
