% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pf_second_order}
\alias{pf_second_order}
\title{Calculate a 'second order' \code{pfc}, which represents summed
branch length from each node to the terminal node along each flow.
When used in modelling traits, a second order \code{pfc} implies that
rates of evolution of the trait are themselves evolving according to
Brownian motion.}
\usage{
pf_second_order(x, ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{...}{Additional arguments passed to or other methods.}
}
\value{
A \code{pfc} object
}
\description{
Calculate a 'second order' \code{pfc}, which represents summed
branch length from each node to the terminal node along each flow.
When used in modelling traits, a second order \code{pfc} implies that
rates of evolution of the trait are themselves evolving according to
Brownian motion.
}
\examples{
pf_second_order(rpfc(100))
}
