% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_standardise}
\alias{pf_standardise}
\title{Standardise the phylogenetic flow features to an implied typical
variance of 1.}
\usage{
pf_standardise(x)
}
\arguments{
\item{x}{a \code{pfc} object}
}
\value{
a scaled \code{pfc} object
}
\description{
Can be used to make random effects based on \code{pfc}s comparable.
}
\examples{
pf_standardise(rpfc(100))
}
