% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_tips}
\alias{pf_tips}
\title{Get only the tip elements of a \code{pfc}}
\usage{
pf_tips(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A \code{pfc} object with only tip elements
}
\description{
Get only the tip elements of a \code{pfc}
}
\examples{
pf_tips(rpfc(100))
}
