% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pf_vcv}
\alias{pf_vcv}
\title{Calculate phylogenetic variance covariance matrix from \code{pfc}}
\usage{
pf_vcv(x, ...)
}
\arguments{
\item{x}{A \code{pfc} object}

\item{...}{Other arguments for future extensions.}
}
\value{
A \code{sparseMatrix}
}
\description{
Calculate phylogenetic variance covariance matrix from \code{pfc}
}
\examples{
pf_vcv(rpfc(10))
}
