% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_zeros}
\alias{pf_zeros}
\title{Replace features with zeros}
\usage{
pf_zeros(x)
}
\arguments{
\item{x}{A \code{pfc} object}
}
\value{
A \code{pfc} object with all feature values replaced with zeros
}
\description{
Replace features with zeros
}
\examples{
pf_zeros(rpfc(100))
}
