% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pfc_from_pftibble}
\alias{pfc_from_pftibble}
\title{Convert pftibble into a pfc}
\usage{
pfc_from_pftibble(pft)
}
\arguments{
\item{x}{a pftibble \code{tibble}}
}
\value{
A \code{pfc}
}
\description{
Convert pftibble into a pfc
}
\examples{
pfc_from_pftibble(pf_as_pftibble(rpfc(100)))
}
