% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-package.R
\docType{package}
\name{phyf-package}
\alias{phyf}
\alias{phyf-package}
\title{phyf: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
