% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primate_diet}
\alias{primate_diet}
\title{Primate Diet Diversity and Threat Status Data with Phylogeny}
\format{
\subsection{\code{primate_diet}}{

A 'pf' data frame (subclasses \code{tibble}) with 504 rows and 55 columns:
\describe{
\item{label}{Node labels including species name for the tip labels}
\item{phlo}{The phylogenetic flow column which stores the phylogenetic information}
\item{Threat status}{IUCN threat category}
\item{Threat status source}{Threat status data source -- see \code{primate_refs} dataset}
\item{Body mass (g)}{Primate species' mean body mass in grams}
\item{Body mass source)}{Body mass data source -- see \code{primate_refs} dataset}
\item{Range size (km2)}{Primate species' range size in kilometers}
\item{Range size source)}{Range size data source -- see \code{primate_refs} dataset}
\item{Diet disparity (PSV)}{Primate species' disparity of diet items as
measured by Phylogenetic Species Variability (PSV) metric.}
\item{Diet breadth}{Primate species' breadth of diet items as measured by
the number of different diet items (richness).}
\item{Diet diversity (DDI)}{Primate species' diversity of diet items as measured by
the DDI metric.}
\item{Trophic guild}{Primate species' trophic guild. Possible values:
"Omnivore", "Frugivore", "Gummivore", "Insectivore", or "Folivore-frugivore"}
\item{diet_item:item (40 columns)}{Each of the 40 columns starting with "diet_item:"
represents a different type of item in primates' diets. These are binary integer
columns with a 1 if the species feeds on that diet item or 0 if it does not.}
}
}
}
\source{
\url{https://zslpublications.onlinelibrary.wiley.com/doi/full/10.1111/acv.12823}
}
\usage{
primate_diet
}
\description{
Data on primates diets and their threat status joined to a \code{pf} object (for \code{{phyf}})
}
\references{
Machado, F. F., Jardim, L., Dinnage, R., Brito, D., & Cardillo, M. (2022). Diet disparity and diversity predict extinction risk in primates. Animal Conservation.
}
\keyword{datasets}
