% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primate_diet_refs}
\alias{primate_diet_refs}
\title{Primate Diet Data References}
\format{
\subsection{\code{primate_diet_refs}}{

A 'pf' data frame (subclasses \code{tibble}) with 4 rows and 3 columns:
\describe{
\item{FOOD ITEM}{Food item name. This matches the 'item' 'diet_item:item'
columns in  the \code{primate_diet} data set}
\item{level 2}{A category categorizing the food items immediately above
the items themselves}
\item{level 3:level 6 (4 columns)}{Additional categories arranged in an
heirarchy}
}
}
}
\source{
\url{https://zslpublications.onlinelibrary.wiley.com/doi/full/10.1111/acv.12823}
}
\usage{
primate_diet_refs
}
\description{
A \code{data.frame} with references for the Primate Diet Data. See the \code{primate_diet}
data set.
}
\references{
Machado, F. F., Jardim, L., Dinnage, R., Brito, D., & Cardillo, M. (2022). Diet disparity and diversity predict extinction risk in primates. Animal Conservation.
}
\keyword{datasets}
