% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{rpfc}
\alias{rpfc}
\title{Generate a random tree and return it as a
\code{pfc} object}
\usage{
rpfc(n, method = ape::rcoal, ...)
}
\arguments{
\item{n}{Number of tips in the generated tree.}

\item{method}{A function to generate a tree. Default
is \code{ape::rcoal}. See \code{ape::rtree()} for more options.}

\item{...}{Additional arguments to pass to \code{method}}
}
\value{
a \code{pfc} object
}
\description{
Generate a random tree and return it as a
\code{pfc} object
}
\examples{
plot(rpfc(100))
}
