% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vert_bmr}
\alias{vert_bmr}
\title{Vertebrate Base Metabolic Rates with Phylogeny}
\format{
\subsection{\code{vert_bmr}}{

A 'pf' data frame (subclasses \code{tibble}) with 1,712 rows and 8 columns:
\describe{
\item{label}{Node labels including species name for the tip labels}
\item{phlo}{The phylogenetic flow column which stores the phylogenetic information}
\item{lnBMR}{Natural log of the base metabolic rate}
\item{lnMass}{Natural log of body mass}
\item{lnMass2}{Squared natural log of body mass}
\item{lnGS}{Natural log of genome size}
\item{endo}{Is the species endothermic? 1 for yes, 0 for no}
}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.3c6d2}
}
\usage{
vert_bmr
}
\description{
Data on vertebrate base Metabolic rates joined to a \code{pf} object (for \code{{phyf}})
}
\references{
Uyeda JC, Pennell MW, Miller ET, Maia R, McClain CR. The evolution of energetic scaling across the vertebrate tree of life. The American Naturalist. 2017 Aug 1;190(2):185-99.
}
\keyword{datasets}
