% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pf_as_pfc.phylo}
\alias{pf_as_pfc.phylo}
\title{Make a \code{pfc} object from a \code{xlo} object}
\usage{
\method{pf_as_pfc}{phylo}(x, ...)
}
\arguments{
\item{x}{A phylogenetic tree in \code{ape::phylo}
format}
}
\value{
a \code{pfc}
}
\description{
Make a \code{pfc} object from a \code{xlo} object
}
\examples{
pf_as_pfc(ape::rtree(100))
}
