% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{pf_filter_with_mrca}
\alias{pf_filter_with_mrca}
\title{Filter descendents of a MRCA}
\usage{
pf_filter_with_mrca(
  x,
  descendents,
  drop_null_edges = TRUE,
  drop_root_edges = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{pf} or \code{pfc} object}

\item{descendents}{An index into a \code{pfc}, using any way
that you can index a \code{pfc}. \code{filter_with_mrca} will find
the most recent common ancestor of everyhting in
\code{descendents} and then it will return a filtered \code{pf}
or \code{pfc} with all descendents of the mrca.}

\item{drop_null_edges}{If \code{TRUE} (the default), drop edges with
no descendants after the filtering.}

\item{drop_root_edges}{If \code{TRUE} (the default), drop all edges
leading up to the mrca (e.g. when combined together these would make
a root edge leading to the desired clade)}

\item{...}{Other arguments for future extensions.}
}
\value{
A \code{pf} or \code{pfc} object
}
\description{
Filter a \code{pf} or \code{pfc} object by taking all
descendents of the most recent common ancestor (mrca)
of \code{descendents}
}
\examples{
avonet \%>\% 
    pf_filter_with_mrca(label \%in\% c("Platalea_minor", "Pelecanus_occidentalis"))
}
