% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pfc}
\alias{pfc}
\title{Create a new phylogenetic flow collection object (\code{pfc})}
\usage{
pfc(
  pfn = character(),
  pfpp = pfp(),
  pfl = list(),
  is_tip = logical(),
  edge_names = character(),
  edge_lengths = double(),
  sparse_mat = NULL
)
}
\arguments{
\item{x}{\itemize{
\item For \code{pfc()}: A phylogenetic tree in \code{ape::phylo} format
\item For \code{is_pfc()}: An object to test.
}}
}
\description{
Create a new phylogenetic flow collection object (\code{pfc})
}
