% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-vctrs.R
\name{pfp}
\alias{pfp}
\alias{pf_is_pfp}
\alias{pf_is_pfc}
\title{Create a new phylogenetic flow path object (\code{pfp})}
\usage{
pfp(x = list())

pf_is_pfp(x)

pf_is_pfc(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{pfp()}: A list of nodes (integer) the path passes through
\item For \code{is_pfp()}: An object to test.
}}
}
\value{
A \code{pfp} object
}
\description{
Create a new phylogenetic flow path object (\code{pfp})
}
\examples{
pfp(ape::nodepath(ape::rtree(100)))
}
