% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyf-package.R
\docType{package}
\name{phyf-package}
\alias{phyf}
\alias{phyf-package}
\title{phyf: Phylogenetic Flow Objects for Easy Manipulation and Modelling of Data on Phylogenetic Trees and Graphs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The {phyf} package implements a tibble and vctrs based object for storing phylogenetic trees along with data. It is fast and flexible and directly produces data structures useful for phylogenetic modelling in the {fibre} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rdinnager.github.io/phyf/}
}

}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/0000-0003-0846-2819}{ORCID})

}
\keyword{internal}
