% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.pf}
\alias{plot.pf}
\title{Make a plot for a \code{pf} object}
\usage{
\method{plot}{pf}(x, columns = NULL, layout = "fan", ...)
}
\arguments{
\item{x}{A \code{pf} obect to plot}

\item{columns}{Bare column names of variables to plot with tree.}

\item{layout}{Either 'phylogram' or 'fan'}

\item{...}{Other arguments passed to \code{phytools::contMap()}}
}
\value{
A \code{phytools::contMap()} object.
}
\description{
Make a plot for a \code{pf} object
}
